+++ NEWS.md
@@ -5,13 +5,14 @@
 * `sjt.itemanalysis()` now works on ordered factors. A clearer error message was added when unordered factors are used. The old error message was not helpful.
 * `plot_likert()` showed category labels in the top and bottom legends in two rows if there are more than six categories. Also, the categories are ordered column wise instead of row wise. This behaviour can now be controlled for grouped likert plots, using `group.legend.options`. The ordering now defaults to row wise and the user can force all categories onto a single row.
 * `tab_model()` gets a `show.aicc`-argument to show the second order AIC.
-* `view_df()` now automatically drops unused value labels from data, so output of frequencies and value labels matches.
+* `tab_model()` gets a `show.reflvl`-argument to show the reference level of factors.
 
 ## Bug fixes
 
 * Argument `string.est` in `tab_model()` did not overwrite the default label for the estimate-column-header.
 * Minor fix in `tab_model()` for mixed models that can't compute R2.
 * The `plot_likert()` option `reverse.scale = TRUE` resulted in `values = "sum.inside"` being outside and the other way around. This is fixed now. 
+* `view_df()` mixed up labels and frequency values when value labels were present, but no such values were in the data.
 
 # sjPlot 2.6.3